Passes tests but uses u32 instead of u64

+++ bindings/src/preamble.ts
@@ -4,14 +4,14 @@ import { JSONEncoder } from "assemblyscript-json";
 import { JSONDecoder, ThrowingJSONHandler, DecoderState } from "assemblyscript-json";
 // Runtime functions
 @external("env", "read_register")
-declare function read_register(register_id: u64, ptr: u64): void;
+declare function read_register(register_id: u32, ptr: u32): void;
 @external("env", "register_len")
-declare function register_len(register_id: u64): u64;
+declare function register_len(register_id: u32): u32;
 
 @external("env", "input")
-declare function input(register_id: u64): void;
+declare function input(register_id: u32): void;
 @external("env", "value_return")
-declare function value_return(value_len: u64, value_ptr: u64): void;
+declare function value_return(value_len: u32, value_ptr: u32): void;
 @external("env", "panic")
 declare function panic(): void;
 `;
